third_party/WebKit/LayoutTests/http/tests/notifications/serviceworkerregistration-service-worker-image-redirect-loop.html

<!doctype html>
<html>
  <head>
    <title>Notifications: Showing a notification with an image that has a redirect loop still succeeds.</title>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
    <script src="../serviceworker/resources/test-helpers.js"></script>
    <script src="resources/test-helpers.js"></script>
  </head>
  <body>
    <script>
      // Tests that showing a notification with an image that has a redirect loop still succeeds.

      async_test(function(test) {
          var scope = 'resources/scope/' + location.pathname,
              script = 'instrumentation-service-worker.js';

          testRunner.setPermission('notifications', 'granted', location.origin, location.origin);
          getActiveServiceWorkerWithMessagePort(test, script, scope).then(function(info) {
              // (1) Display a Web Notification through the Service Worker with
              // an image that takes redirects to another image.
              return sendCommand(info.port, {
                  command: 'show',

                  title: 'My Notification',
                  options: { body: 'Hello, world!',
                             icon: '/resources/infinite-loop.php' }
              });
          }).then(function(data) {
              // (2) Confirm that the service worker displayed the notification successfully.
              assert_true(data.success);

              test.done();
          }).catch(unreached_rejection(test));

      }, 'Showing a notification with an image that has a redirect loop still succeeds.');
    </script>
  </body>
</html>
